+++ server/routes/authRoute.js
@@ -1,13 +1,7 @@
 import { Router } from "express";
-import { syncUserWithDB, getUserWithDB } from "../controller/authController.js";
+import { getUserWithDB } from "../controller/authController.js";
 
 export const userRouter = Router();
 
-userRouter.get("/sync-user", syncUserWithDB);
+// userRouter.get("/sync-user", syncUserWithDB);
 userRouter.get("/alluser", getUserWithDB);
-
-// userRouter.get("/sync-user",requireAuth(), syncUserWithDB);
-// route.get();
-// route.delete();
-// route.patch();
-// route.post();
