+++ server/routes/cardRoute.js
@@ -1,0 +1,5 @@
+import { Router } from "express";
+import { createCard } from "../controller/cardController.js";
+import { clerkClient, requireAuth, getAuth } from '@clerk/express'
+export const cardRouter = Router();
+cardRouter.post("/create-card",  requireAuth(),createCard)
