+++ server/routes/syncRoute.js
@@ -1,0 +1,4 @@
+import { Router } from "express";
+import { syncUserWithDB } from "../controller/syncController.js";
+export const syncRouter = Router();
+syncRouter.get("/user-synced", syncUserWithDB);
